0.2.1

See merge request !57

+++ pojo/src/main/java/com/platon/browser/dto/StatisticInfo.java
@@ -12,9 +12,9 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
 public class StatisticInfo {
     private long avgTime;
 
-    private double current;
+    private long current;
 
-    private double maxTps;
+    private long maxTps;
 
     private BigDecimal avgTransaction;
